requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
python-dotenv~=1.0.1
psycopg2~=2.9.9
requests~=2.32.3
aiogram~=2.14
numpy~=2.0.0
pillow~=10.4.0
pip~=24.1.2
attrs~=23.2.0
pytz~=2024.1
PyYAML~=6.0.1
idna~=3.7
multidict~=6.0.5
Babel~=2.15.0
setuptools~=71.1.0
asgiref~=3.8.1
sqlparse~=0.5.0
aiohttp~=3.9.5
certifi~=2024.7.4
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
urllib3~=2.2.2
packaging~=24.1
inflection~=0.5.1
uritemplate~=4.1.1
google-api-python-client~=2.137.0
Markdown~=3.6
PyJWT~=2.8.0